manifold.py: export 'RetrAsExpMixin'
