+++ pymanopt/manifolds/manifold.py
@@ -1,7 +1,21 @@
 import abc
 import functools
+import warnings
 
 import numpy as np
+
+
+class RetrAsExpMixin:
+    """Mixin class which defers calls to the exponential map to the retraction
+    and issues a warning.
+    """
+
+    def exp(self, Y, U):
+        warnings.warn(
+            "Exponential map for manifold '{:s}' not implemented yet. Using "
+            "retraction instead.".format(self.__class__.__name__),
+            RuntimeWarning)
+        return self.retr(Y, U)
 
 
 class Manifold(metaclass=abc.ABCMeta):
