+++ pymanopt/manifolds/psd.py
@@ -1,5 +1,3 @@
-import warnings
-
 import numpy as np
 from numpy import linalg as la, random as rnd
 from scipy.linalg import expm
@@ -9,21 +7,9 @@ try:
 except ImportError:
     from scipy.linalg import solve_lyapunov as lyap
 
-from pymanopt.manifolds.manifold import EuclideanEmbeddedSubmanifold, Manifold
+from pymanopt.manifolds.manifold import (EuclideanEmbeddedSubmanifold,
+                                         Manifold, RetrAsExpMixin)
 from pymanopt.tools.multi import multilog, multiprod, multisym, multitransp
-
-
-class _RetrAsExpMixin:
-    """Mixin class which defers calls to the exponential map to the retraction
-    and issues a warning.
-    """
-
-    def exp(self, Y, U):
-        warnings.warn(
-            "Exponential map for manifold '{:s}' not implemented yet. Using "
-            "retraction instead.".format(self._get_class_name()),
-            RuntimeWarning)
-        return self.retr(Y, U)
 
 
 class SymmetricPositiveDefinite(EuclideanEmbeddedSubmanifold):
@@ -154,7 +140,7 @@ class SymmetricPositiveDefinite(EuclideanEmbeddedSubmanifold):
 #              psd matrices, or in fixed_rank. Alternatively, move this one and
 #              the next class to a dedicated 'psd_fixed_rank' module.
 
-class _PSDFixedRank(Manifold, _RetrAsExpMixin):
+class _PSDFixedRank(Manifold, RetrAsExpMixin):
     def __init__(self, n, k, name, dimension):
         self._n = n
         self._k = k
@@ -293,7 +279,7 @@ class PSDFixedRankComplex(_PSDFixedRank):
         return rand_() + 1j * rand_()
 
 
-class Elliptope(Manifold, _RetrAsExpMixin):
+class Elliptope(Manifold, RetrAsExpMixin):
     """
     Manifold of n-by-n psd matrices of rank k with unit diagonal elements.
